single adapter handling different views
