+++ app/src/main/java/co/infinum/productive/fragments/ProjectsFragment.java
@@ -21,7 +21,6 @@ import butterknife.ButterKnife;
 import co.infinum.productive.R;
 import co.infinum.productive.activities.TasksListActivity;
 import co.infinum.productive.adapters.ProjectAdapter;
-import co.infinum.productive.adapters.ProjectSectionAdapter;
 import co.infinum.productive.dagger.components.DaggerProjectsComponent;
 import co.infinum.productive.dagger.modules.ProjectsFragmentModule;
 import co.infinum.productive.listeners.OnProjectClickListener;
@@ -48,9 +47,7 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
     @Bind(R.id.empty_projects_info)
     TextView emptyProjectsInfo;
 
-    private RecyclerView.Adapter mAdapter;
-
-    private ProjectSectionAdapter mSectionAdapter;
+    private ProjectAdapter mSectionAdapter;
 
     private Context context;
 
@@ -95,9 +92,9 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
         }
 
         if (isRefreshed) {
-            refreshAdapters(projectTiles);
+            refreshAdapter(projectTiles);
         } else {
-            initAdapters(projectTiles);
+            initAdapter(projectTiles);
         }
     }
 
@@ -118,11 +115,8 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
         mRecyclerView.setLayoutManager(layoutManager);
     }
 
-    private void initAdapters(ArrayList<ProjectTile> projectTiles) {
-        mAdapter = new ProjectAdapter(context, getResources(), projectTiles, this);
-        mSectionAdapter = new ProjectSectionAdapter(context, R.layout.projects_list_item_separator, mAdapter);
-
-        ((ProjectAdapter) mAdapter).setProjectSectionAdapter(mSectionAdapter);
+    private void initAdapter(ArrayList<ProjectTile> projectTiles) {
+        mSectionAdapter = new ProjectAdapter(context, getResources(), projectTiles, this);
 
         setSections(projectTiles);
 
@@ -130,7 +124,7 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
     }
 
     private void setSections(ArrayList<ProjectTile> projectTiles) {
-        ArrayList<ProjectSectionAdapter.Section> sections = new ArrayList<>();
+        ArrayList<ProjectAdapter.Section> sections = new ArrayList<>();
 
         // calculates offset for each client and sets sections
         String prevClientName = "";
@@ -144,24 +138,24 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
                 ++currentOffset;
             } else {
                 overallOffset += currentOffset;
-                sections.add(new ProjectSectionAdapter.Section(overallOffset, currClientName));
+                sections.add(new ProjectAdapter.Section(overallOffset, currClientName));
                 currentOffset = 1;
             }
 
             prevClientName = currClientName;
         }
 
-        ProjectSectionAdapter.Section[] sectionsList = new ProjectSectionAdapter.Section[sections.size()];
+        ProjectAdapter.Section[] sectionsList = new ProjectAdapter.Section[sections.size()];
 
         mSectionAdapter.setSections(sections.toArray(sectionsList));
     }
 
-    private void refreshAdapters(ArrayList<ProjectTile> projectTiles) {
-        if (mAdapter != null && mSectionAdapter != null) {
+    private void refreshAdapter(ArrayList<ProjectTile> projectTiles) {
+        if (mSectionAdapter != null) {
             isRefreshed = false;
 
             setSections(projectTiles);
-            ((ProjectAdapter) mAdapter).refresh(projectTiles);
+            mSectionAdapter.refresh(projectTiles);
 
             swipeRefreshLayout.setRefreshing(false);
         }
@@ -170,7 +164,7 @@ public class ProjectsFragment extends BaseFragment implements ProjectView, OnPro
     @Override
     public void onProjectsClick(int position) {
         Intent intent = new Intent(getActivity(), TasksListActivity.class);
-        intent.putExtra(PROJECT, ((ProjectAdapter) mAdapter).getItem(position));
+        intent.putExtra(PROJECT, mSectionAdapter.getItem(position));
         startActivity(intent);
     }
 }
